Renaming OneLakeClient to OneLakeStorageClient
